server/requirements.txt
g4f
pywebview
flask
flask_cors
